Remove references to TimeWarp.Statiq will extract common stuff later
Change Author.Url to method Author.GetUrl()
Add PathUtlities.CleanPath to put this functionality in one place
Chaged Unicode format of the Thai article and yet still it isn't rendering correctly.,
Removing the diagnostics parials as I dont' need diagnostics right now. And this is a security leak if left on.